requirements.txt
# Testing
pytest>=7.0.0

# Gmail API
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# PrintNode API
requests>=2.31.0

# CSV and data handling
pandas>=2.0.0

# Scheduling
schedule>=1.2.0

# Environment variables
python-dotenv>=1.0.0
